+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## v5.0.0 2019-07-15
 
 - Uses promise-readable and promise-writable v5.
+- `PromiseReadable` implements `AsyncIterable` so it is possible to use
+  `for await (const chunk of promiseReadable)` loop.
+- New method `iterate` is provided.
 
 ## v4.0.3 2019-06-04